Swap axes for all scalar arrays
Unfortunately, vtkImagePermute only swaps the axes of the active
scalars. This causes some big problems later if there are multiple
scalar arrays in the image data.

This commit adds a unified function for swapping the axes that will
swap axes of all of the scalars. The parts of the code that need to
swap axes of the image data should call this function.

Additionally, this also adds a common function for reading in
the angles in GenericHDF5Format. The various HDF5 formats that need
to read in angles now call this function.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>