Print docker logs as they come
This adds a command to the docker executor, "docker logs -f <id>",
which prints the stdout and stderr from the container as it comes
out. The docker executor then prints the stdout and stderr to the
console and the tomviz "Messages", so that the user can see the
docker output.

This also sets the environment variable PYTHONUNBUFFERED in the
tomviz/pipeline docker image, so that the python output gets printed
immediately rather than when the execution finishes.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>