Make child dataset for external match internal
Previously, the child dataset for the external pipeline would create
a deep copy of the parent dataset. This was not done in the internal
dataset. Instead, in the internal dataset, the child dataset would be
empty, and it would only have spacing and dimensions that matched that
of the parent (with the exception of when the parent was a tilt series,
then the spacing of the tilt axis of the child would be altered).

Additionally, in the internal dataset, if the active_scalars were set
on the empty child dataset, a new scalars named 'Scalars' would be
created.

The changes in this commit cause the external dataset to match the
behavior of the internal dataset, in that, the child dataset is an
empty dataset with the same spacing as that of the parent (except for
the tilt series parent case), and if the active scalars get set on an
empty child dataset, a new scalars named 'Scalars' would be created.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>