Allow non-active scalars to be modified in python
In the python operator "Dataset" classes, allow the non-active
scalars to be modified via "set_scalars()". If there are scalars
already present in the dataset, and the newly set scalars have different
dimensions compared to the other scalars, the other scalars are
discarded.

Additionally, this allows a child dataset to be either a volume
or another tilt series, so that the user can indicate whether the tilt
angles and spacing should be preserved.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>