Give the tilt axis a spacing of 1.0
In the EMD format, the spacing of the tilt axis was previously
calculated by comparing the distance between the first and second
angles. Other formats, though, like the GenericHDF5 format, do not
give the tilt axis a spacing.

For consistency, and so that the data does not change spacing when
a GenericHDF5 file is loaded and used in an external pipeline, use
a spacing of 1.0 both in the EMD format, and in the pipeline
executor.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>