Add ability to color contours by other scalars
This adds a checkbox to the contour panel, "Color by Array",
which, when checked, colors the contour by the array selected in
the combo box next to it. It allows the user to color the contour
by scalars other than the active one.

Internally, it uses a vtkProbeFilter to map the voxel values of the
other array onto the contour. This results in it being a little
slower to render, and a little lower resolution contour.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>