Rebuild paraview CI for out-of-date dependencies
When python, Qt, or the paraview SHA change versions or get
upgraded, automatically re-build ParaView. This will hopefully
help fix some of the issues of builds breaking often because of
dependency changes.

We should be able to easily add more dependencies to the checker
to cause ParaView to re-build, if we find other dependency
upgrades are causing issues.

This also upgrades the CacheBeta@0 task in azure-pipelines to its
non-beta version, Cache@2.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>