Add functions to save/load tvh5 file
The tvh5 file is a full state file, in that the tomviz state is saved,
and all of the data sources are saved as well.

The full state file is kind of like an EMD file. It has an EMD node at
/data/tomography, which is where the active data source is saved. But
it is extended so that at the root of the EMD file, there is a
"tomviz_state" dataset, which contains the tomviz state, and there is
a "tomviz_datasources" group, which contains all of the data sources.
Each data source is saved as an EMD node, and the group each data
source is in is the id of the data source (which matches the id of the
data source in the state file). The data source that is active is
created via a soft link to /data/tomography, so it is not created
twice.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>