Add function to set views with a state array
This was previously being done in ModuleManager::deserialize() in order
to restore the previous state of the views after deserializing data
sources, because the views get modified during data source
deserialization.

Instead of only performing it in ModuleManager::deserialize(), extract
the code into a function, and we can now use it in the full state
files.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>