Add "Save State" action for most recent state file
The "Save State" action gets enabled if there is a most recent
state file set. The most recent state file gets set in two different
conditions:

1. The user chooses a file to save via the "Save State As" action with
   its file dialog.
2. A state file gets loaded, whether done programmatically or via a
   file dialog.

Because the SaveLoadStateReaction functions are mostly static, we
needed to be able to set the most recent state file on a singleton.
The ModuleManager is being used for this purpose.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>