Enable active scalar selection for the ModuleSlice
This PR brings the slice to feature parity with the ModuleVolume.
For a multi-scalar datasource, the active scalars shown by the slice can either
be in synch with what is selected upstream on the datasource, or they
can be locked to a specific array. The user can freely go back and
forth.

At the implementation level, I created a subclass of the
vtkTrivialProducer, the vtkActiveScalarsProducer.
vtkActiveScalarsProducer stores internally a shallow copy of the
vtkImageData that is set as output. The active scalars can be set on the
shallow copy with no repercussion on the original data.

Signed-off-by: Alessandro Genova <alessandro.genova@kitware.com>