Allow tvh5 files to be read from tvsm
This requires that the data source has set on it the path to
the node in the tvh5 file from which the data source was
generated. This path gets saved into the tvsm file, and if
a .tvh5 file is read in as data rather than as a state file,
it requires a path to be specified to generate the data source.

This also allows the autosave file to work.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>