Allow clipping of slice modules
The slice module uses clipping planes to trim the TexturePlaneActor and keep it
within the outline bounds, but a mapper can only have a maximum of 6 planes. To
give clipping planes from from modules prority they are always added first. The
actor bounds are then checked and clipped if needed, until the maximum number of
planes has been reached.

Signed-off-by: Brianna Major <brianna.major@kitware.com>