Use item data for storing QComboBox string options
We can't be certain that the QComboBox text will be unchanged. For
instance, ampersands can be added, or the text can be translated. So
it is better to not use the text from the QComboBox text.

Instead, store the text in the item data, where it should not be changed,
and use that instead.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>