Use proxies to communicate from Python to C++
The PyBind11 classes to interact with operators and the pipeline provide
a nice, simple interface for Python to interact with the C++
application. In order to more fully decouple the two a core library was
added that uses empty base classes. These are then implemented in the
main Tomviz code, and the appropriate factories are set so that the
running instance can interact with the pipeline.

This enables the Python wrapping to only connect to tomvizcore, and the
relevant VTK classes. Appropriate export macros need to be added,
tested, and verified on a much smaller API surface. The rest of the code
can be left alone, and new proxies added in future to wrap limited
portions of the API that should be exposed to Python.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>