Ensure array is active in DataSource constructor
Some of the filters from VTK/ParaView don't always set an active
array. This was the case for using the PythonCalculator to merge
arrays together.

To make sure an array is active, check to see if there are arrays
and that no array is active, and if that is the case, make one
active.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>