Get actual range of all components for histogram
arrayPtr->GetFiniteRange(minmax, -1) was being used before, and this
actually returns the range of the L2 norm over all of the components.
This was resulting in some negative indices being computed in the
histogram computation code, and a subsequent seg fault.

Instead, loop over all the components and return the actual range
of all the components. This fixes the segmentation fault.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>