Post process jsonpatch so we can use upstream package
Post process the patch produced by jsonpatch to it conforms to
Tomviz semantics. The semantic for Tomviz state is a little
different than a simple dict. You can't change a operator or
module by simply patch id of a dict. If an id is different
the the objects should be treated as immutable. So we need
to post process the operations and replace any attempts to
patch object ids with a operation to replace the whole
object.

Signed-off-by: Chris Harris <chris.harris@kitware.com>