Map 3-component data to RGBA values for volumes
There are some things here that can be improved, but this is a basic
working example.

This takes the 3-components and and rescales their range to be 0 - 1 (so
the full range of RGB is used). Then, it places the norm of the original
3-components into the fourth component (which gets passed through the
opacity functions to determine opacity). This allows utilizing the
opacity editor to modify the opacity of the image.

The histogram appears to be displaying the norm of the components of
the data.

Internally, we are using vtkVolumeProperty's IndependentComponents, which
requires 2 or 4-component data. For 4-component data, the first three
components are RGB, and the fourth component is passed through the opacity
functions for opacity.

An option hasn't been added yet to turn this mapping off, so currently, all
3-component data maps to RGBA values for volumes.

Also, this only works for float/double data. Needs to be generalized more
for integers.

One possible improvement is that a filter might be better instead of an
internal object for mapping 3-component data to 4-component RGBA data.

It might also be good to make sure the opacities are working properly...
it seems to me that the opacities of this are off.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>