Wrap volume mapper with vtkSmartVolumeMapper
vtkGPUVolumeRayCastMapper is still being used internally inside
vtkSmartVolumeMapper, but it is now wrapped with vtkSmartVolumeMapper to
allow us to utilizes the features of vtkSmartVolumeMapper. In particular,
we may try to use the vector mode features of vtkSmartVolumeMapper.

vtkSmartVolumeMapper was also subclassed to give us access to
more parts of vtkGPUVolumeRayCastMapper.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>