Update client view geometry bounds on slice change
I do not know why this is not being updated automatically. It is possible
there is a better fix out there that gets the vtkNonOrthoImagePlaneWidget
to get the client view geometry bounds to be updated automatically. But I
am not immediately aware of one.

As such, force the geometry bounds to be updated in the slice module
every time a slice changes.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>