Add image viewer mode to the view menu
This sets up a slice module, moves the camera, and adds a slider widget
below the central widget so that the user can slide through the
images/slices.

When checked, these things are done:

1. Sets the projection to orthographic
2. Sets the render interactor mode to 2D
3. Finds the first slice module or creates one. Hides all other modules.
4. Align the camera so the single slice is filling the screen
5. Show the slider widget below the central widget

This assumes that the images are in the Z-direction as well.

When the user unchecks the image viewer mode, the projection type,
render interactor, modules, and most of the view are restored to
the settings they had before.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>