Re-build LUT every time it is modified
This gives us live updates for edits to the color bar. So now, things
like deleting or moving a control point will give live updates, where
they did not give live updates before (and you had to do things like
creating a new control point or inverting the color map to get it to
update).

Strangely, re-building the LUT is also considered a modify event, so
we have to block the signal to avoid infinite recursion.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>