Add a brightness and contrast editor
This allows the user to edit the min/max, and brightness/contrast of
the LUT via some sliders. The LUT is rescaled to match the ranges
specified by the settings.

Placeholder nodes are kept on the sides to make it easier to see the
colors, and to allow editing of the full color bar.

The brightness is essentially the average of the min and max. The
contrast is a measure of the distance between the min and max
relative to the min and max of the data.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>