Added basic FXI workflow operator and widget
This allows the user to perform test rotations by running some python
code in the python script (specifically, it looks for and runs the
"test_rotations()" function in the script). It then displays the data
in the widget with a slider, so the user can see which rotation to use
for the reconstruction.

Most things are working properly. There are some possible improvements
available, though.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>