Always default to data bounds for slice range
We do indeed read in the settings, which may have been set to some
other range, but having to look up the data range and set it can
be an unpleasant extra step. Instead, set the min/max to be the
data bounds as a default.

Also, communicate in the tooltip what the max is.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>