Allow custom parameters in FXI workflow operator
This commit adds the ability for the user to add custom parameters
to both the operator UI and the main `transform()` function via the
`description.json` file for the FXI workflow operator.

There is now a general `setupUI()` function on the
`CustomPythonOperatorWidget` that matches the signature of `setupUI()`
in the `OperatorWidget` class. The base class version does not do
anything, but the FxiWorkflowWidget will use it to check for extra
parameters and use the `InterfaceBuilder` class to add them to a
dedicated QLayout. These additional parameters are utilized in
`getValues()` and `setValues()`.

An additional `updateWidgetValues()` function was added to the
`InterfaceBuilder` that takes a `QObject` similar to `parameterValues()`,
which updates the widget values to match the values in m_parameterValues
(which can be set via `InterfaceBuilder::setParameterValues()`).

Some of the functionality in `FxiWorkflowWidget::setupUI()` may be better
off transfered to `CustomPythonOperatorWidget::setupUI()` to make it
easier to do this kind of thing in other operators in the future. Currently,
though, it is all in `FxiWorfklowWidget`.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>