Add ability to set custom axes titles to outline
This adds a "Custom Axes Titles" checkbox to the outline module
properties. If checked, a new group box will appear that allows the
user to set the axes titles via line edits.

The custom axes title settings are serialized and deserialized, so
they can be saved to state files as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>