Add "tag" key to description.json parameters
When the "tag" key is present and not an empty string, that indicates
the parameter is supposed to be treated differently than the others
parameters. The default interface builder will currently just skip
over any parameters which have a non-empty tag.

Python operators with custom widgets can use these tags for special
parameters, and can call InterfaceBuilder::buildParameterInterface()
with the tag to automatically add widgets for these special parameters
to a layout.

This will be used first in the FxiWorkflowWidget.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>