Allow extra test rotation parameters to be added
This allows extra test rotation parameter widgets to be added via the
`description.json` file and used in the `test_rotations()` function.

A parameter in the `description.json` is indicated to be a test
rotation parameter via a tag of "test_rotations". If this is the tag
for the parameter, it will be ignored for the regular `transform()`
function, and instead only be used in the `test_rotations()` function,
which is called by clicking the "Test Rotations" button in the widget.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>