Add time series steps to DataSource and loading
With these changes, if the user opens multiple files, a dialog will
ask the user if they wish to open up the files as a time series. If
they select yes, all of the files will be loaded, but only the data
source from the first file will be kept, and the image data from the
other files will be added as time series steps to the data source.
The files are currently sorted when they are loaded using a regular
string sort, so the first one in this order will be kept, and the
others will be added in order as time series steps.

This also sets up the animation to have a number of time steps equal
to the number of time steps the user loaded, and it adds a connection
in the DataSource to detect when the time changes. This allows the user to
use the VCR toolbar as well as the animation widget to control the time
step of the current data source.

The time series steps are currently not supported in state files, and
any operators will operate on the current image data, and the output
data source will not have any time series steps.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>