Make adding data sources to the pipeline optional
Previously, `LoadDataReaction::loadData()` would always add the data
sources to the pipeline. For time series, we don't want to do this for
all of the files after the first file, but we also don't want to re-write
all of the logic for loading files.

Thus, make adding the data sources to the pipeline optional.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>