Add dataset parameter option to description.json
This creates a QComboBox with all of the available data sources in tomviz
as options. Whichever one the user selects will be passed into the python
function as a `Dataset` object. This allows operators to take in multiple
data sources as input, but operators may still only output one data source.

In order to allow data sources to be selected by label in the combo box,
if this parameter is used, and if any data sources have the same label,
the duplicates will be renamed with "_2", "_3", etc. suffixes, and warning
printed to the terminal that this has been done. This ensures that the
user's selection is unambiguous.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>