Do not re-label data sources in InterfaceBuilder
Instead, we added a function to module manager:
`ModuleManager::allDataSourcesDepthFirst()`, which returns the
Datasources in the same order that they appear in the pipeline view.

Next, we only create unique names for the combo box (and do not
re-label the Datasources). This is one like "Label", "Label (1)",
"Label (2)", etc. These are in the same order as they appear in the
pipeline view, so the user can hopefully figure out which one
corresponds to which data source.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>