Add data orientation and rotation interactions
This adds an orientation to the data source and rotation interactions
via the data properties panel. This also re-works some of the data properties
panel and move active objects design. In particular, the user can perform
different transformations (translation, rotation, and scaling) all
simultaneously.

The orientation is defined in the same way that VTK defines it for its actors,
where the rotation center is at the origin. This allows us to just call
"SetOrientation" on actors to get them to be oriented properly. However, for
the rotation interactions, the rotations are performed with the rotation
center as the center of the data (which gets converted into the VTK version
of orientation). This allows for more intuitive interaction.

This also re-works the internals of MoveActiveObject to just use vtkTransform
rather than trying to do the math ourselves. This results in a lot of
simplification.

This also adds data origin/orientation to state saving.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>