Add manual manipulation widget
This allows the user to visually change the position, rotation, and scaling
of the data, and then apply these transformations to the underlying voxels.

It uses only numpy for shifting and scaling. If there are any rotations,
however, ITK is needed.

The rotations in the manual manipulation widget are set up to rotate about
the center of the data, as opposed to the rotations in the data properties
panel which mirror VTK and rotate about the origin.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>