Add visual alignment with reference data
If the user selects a reference data source in the manual manipulation widget,
the reference data source will be moved so that its center matches up with the
data source being manipulated. This assists in aligning the data, and is
necessary for when the two data sources do not have matching voxels or spacing.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>