Add resampling to manual manipulation
If the user has checked "Align voxels with reference?" in the
manual manipulation widget, then the dataset will be resampled by the
ratio of the spacing to the reference spacing. This makes it so that
visual changes to the size will be applied correctly to the voxels.

An example video of this may be seen [here](https://drive.google.com/file/d/1zxQix4kIp8mzxsffCqipGJN4WLRCZZ1P/view?usp=sharing).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>