Add "File"->"Open" menu with "Time Series" inside
This commit moves "Open Data" and "Open Stack" into the menu also as
"Data" and "Stack", respectively.

This allows us to preserve the previous behavior for selecting multiple
files with "Open Data", and it adds the extra "Open"->"Time Series"
option so that it is completely clear that a time series is being opened
(and this removes the need to ask the user whether it is a time series
or not).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>