Add time series labels
When a data source contains time series data, a label for the current
time step will be displayed in the render window. This label can be
moved and resized.

Additionally, a new section will be present in the data properties
panel called "Time Series", where the only current option inside it to
show/hide the time series label.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>