Use cmake_path to compute path to python libraries
Rather than assuming that the path from the exe dir to the python
libraries is "exe_dir + "/../" TOMVIZ_PYTHON_INSTALL_DIR", compute the
relative path in cmake and use that instead.

When the exe dir was "Library/bin", and the library dir was
"Library/lib", the previous path that was computed was
"../Library/lib/...", which should have been "../lib/..." instead. This
computation fixes the issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>