Add animation helper dialog with camera orbit
The animation helper dialog adds some convenient tools for creating
animations. The first tool that is present is the ability to create
a camera orbit animation. This uses the current camera settings: the
animation starts at the current camera position, and rotates around
the focal point to the right.

Other animation helper tools will be added soon as well...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>