Add time series and modules to animation helper
This adds tabs where the user can switch between "Camera",
"Time Series", and "Modules". There are animation options for
each of these.

For time series, the option to turn off time series animations is
present, but only applicable if a data source was loaded that contains
a time series.

For modules, the user selects a data source and then an animatable module.
Then, the user selects options and adds an animation for the module.
Currently, only contour modules can be animated.

There are also buttons for clearing animations on the various pages. And
an option was also added to change the number of frames.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>