Save data source argument ids correctly
Previously, these would be saved as "null". But now they are being
converted to a string form of the id.

This string form of the id starts with "0x", and we will most likely
check if the string starts with this to determine whether it is pointing
to a data source or not.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>