Add dependencies to data source loading
This allows us to load state files where one operator selected a
another data source. The other data source must be loaded first
before that operator can be ran.

This finds a data source's dependencies, loads in the dependencies
first, waits for their pipeline to finish if the dependency is
a child data source, and then loads in the data source.

It seems to be working well in the examples I've tried.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>