Add tomviz-pipeline support for data dependencies
Some operators may now specify a separate data source as an argument.
In order to support this in the tomviz-pipeline, we have to load in
that data source before running the operator.

This commit adds such dependency support. It enables state files with
multiple data sources to be used with the tomviz pipeline as well, since
multiple data sources is required if an operator has a data source
dependency.

Now, the user has the ability to specify the data source whose pipeline
will be used. If unspecified, the pipeline tries to determine it
automatically. If there is only one data source with operators, it
assumes that is the pipeline. But if there is more than one data source
with operators, it requires the user to specify which one to use.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>