Refactor for independent tomvizcore
This refactors the structure so that the tomvizcore library will not
need to be linked against tomvizlib. This also moves the Variant class
into the `core` directory where it is built into tomvizcore, tomvizlib,
and the pybind11 modules.

This also generalizes the conversions from basic types into python types
and moves it into the pybind11 directory, where it is built into the
pybind11 modules and tomvizlib.

This setup allows the metadata to be dict-like as `map<string, Variant>`.
The setup seems to be working in my testing.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>