Reorder the data and then relabel X and Z
After a fair bit of debugging it is clear that the Python side is silently
reordering the array to Fortran order before the C++ side sees it. This at
least gives a correct image stack with the projections on XY and the angles
along the z axis. Future work to bypass the reorder on the Python side, and
then drop the C++ side reorder before relabeling.

Signed-off-by: Marcus D. Hanwell <mhanwell@bnl.gov>