Fix threshold module with ParaView updates
It was crashing before, because they re-designed the properties.

Now, there is no "ThresholdBetween" property, and you have to set the
upper and lower threshold explicitly.

This re-design now almost matches the old tomviz design, except:

1. In the old design, there was a delay in updating so that you could
move the sliders more freely, and the image would update. Now, it updates
for every slider tick, which makes it much less interactable.
2. There is no reset button. This is probably okay.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>