Add slider tracking to DoubleSliderWidget
This functions similar to KeyboardTracking, in that an update is not
performed until editing is finished. In this case, an update is only
performed if the slider is released. This defaults to `true`, but is
being turned off for the threshold module so that its behavior matches
what it was in Tomviz 1.10.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>