Fix setting render view background color
The color palette at the top changes the background color globally
for all render views. However, the user can right-click individual
render views and set a custom color for that specific render view.

As of the latest paraview updates, we need to set
"UseColorPaletteForBackground" explicitly to zero to ensure that the
individual render view uses its own background color.

This behavior now matches the previous behavior in Tomviz 1.10.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>