Save/load palette color settings
The background color stuff was re-worked in ParaView, so that now a
"UseColorPaletteForBackground" property is present that indicates whether
the global palette should be used or not for the background color.

We also need to save out the global palette color separately rather than
trying to make the render view's background color match the global palette
color.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>