Don't check ConvertDisplayToPointOnSurface return
This function is *intended* to be used to compute a point on a surface.
However, we have been using it anyways to compute the picked center in
empty space.

Newer versions of paraview fixed the return so that it actually returns
`false` now when it doesn't find a surface. But it still moves the center
to where we want it to be.

Thus, stop checking the return and just use it. It works (for now) for
picking the center.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>