Fix python syntax highlighter
ParaView no longer automatically connects the highlighter, so we
must connect it manually.

Additionally, we have to put it before the text is set so that the
text will be highlighted.

Finally, it appears that the convention now in ParaView is to make
the parent the script, so do that.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>