Remove old CI code and scripts
CircleCI had a large security breach in January. Any old secrets need
to be rotated out.

However, we haven't used CircleCI in a long time. Instead of checking
if old secrets are still in place, just eliminate all of the unused CI
in general so we don't need to worry about it.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>