Improve alignment preview image color ranges
For datasets that had a minimum greater than zero, the preview images
in the alignment widget would often appear saturated. That is because
the actual values in the inscribed circle would become much higher than
the values outside of it (which we should ignore anyways).

Thus, when we compute the range to include for the color range, only
check pixel values inside the inscribed circle (and a certain distance
from the edge as well, since edge pixels appear to drop fast in value).

This makes the preview images look much better.

Fixes: #2279

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>