WIP: initial logic for memory check
This adds the initial logic for performing a memory check.

We will estimate the amount of memory loading a file will take (the
estimation may be different for each file type), compare that to the
available system memory, and present a warning to the user if we may
end up using too much memory.

This still needs some work, and we definitely want to make sure it is
cross-platform.

We may also want a way to disable this check, in case some users receive
the warning often and don't want to see it.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>